Collect FormData on form submit and fix Select/RadioGroup state seeding
Forms now collect all named input values via FormData before dispatching
onSubmit actions, ensuring template interpolation resolves even for
untouched inputs.

collectComponentState now inspects Select children for `selected` and
RadioGroup children for `checked` to seed initial state from defaults.

Extracts bindActions, mapProps, and filterInternalProps from renderer.tsx
into prop-transforms.ts to stay within file size limits.
